code/requirements.txt
aenum>=3.1.11
attrs>=22.1.0
cachetools>=5.2.0
click>=8.1.2
Flask>=2.2.2
google-api-python-client>=2.60.0
google-auth>=2.11.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2
google-cloud-core>=2.3.2
google-cloud-run>=0.4.2
google-cloud-error-reporting>=1.6.1
google-cloud-secret-manager>=2.12.6
oauth2client>=4.1.3